+++ src/main/php/Database.php
@@ -8,6 +8,7 @@
  * @package  stubbles\db
  */
 namespace stubbles\db;
+use stubbles\lang\Sequence;
 /**
  * Convenience access to database data to prevent fiddling with query results.
  *
@@ -89,13 +90,18 @@ class Database
      * @param   array   $values         map of values in case $sql contains a prepared statement
      * @param   int     $fetchMode      optional  the mode to use for fetching the data
      * @param   array   $driverOptions  optional  driver specific arguments
-     * @return  array
+     * @return  \stubbles\lang\Sequence
      */
     public function fetchAll($sql, array $values = [], $fetchMode = null, array $driverOptions = [])
     {
-        return $this->dbConnection->prepare($sql)
-                                  ->execute($values)
-                                  ->fetchAll($fetchMode, $driverOptions);
+        return Sequence::of(
+                new QueryResultIterator(
+                        $this->dbConnection->prepare($sql)
+                                           ->execute($values),
+                        $fetchMode,
+                        $driverOptions
+                )
+        );
     }
 
     /**
@@ -126,7 +132,7 @@ class Database
      * @param   string  $sql          sql query to fetch data with
      * @param   array   $values       map of values in case $sql contains a prepared statement
      * @param   int     $columnIndex  number of column to fetch
-     * @return  string[]
+     * @return  \stubbles\lang\Sequence
      */
     public function fetchColumn($sql, array $values = [], $columnIndex = 0)
     {
@@ -151,6 +157,7 @@ class Database
      * @param  string    $sql       sql query to map results of
      * @param  \Closure  $function  function to apply to each result row
      * @param  array     $values    map of values in case $sql contains a prepared statement
+     * @deprecated  use fetchAll()->map()->values() instead, will be removed with 6.0.0
      */
     public function map($sql, \Closure $function, array $values = [])
     {